+++ dune/stuff/test/common.hh
@@ -53,7 +53,7 @@ typedef Dune::tuple<double, float
 
 
 //! where sleep only counts toward wall time, this wastes actual cpu time
-void busywait(const int ms);
+void busywait(const size_t ms);
 
 
 namespace Dune {
